+++ GameLib.Demo/GameLib.Demo.Wpf/ViewModels/ViewModelBase.cs
@@ -1,0 +1,9 @@
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+using System;
+
+namespace GameLib.Demo.Wpf.ViewModels;
+
+public class ViewModelBase : ObservableObject, IDisposable
+{
+    public virtual void Dispose() { }
+}