+++ GameLib.Demo/GameLib.Demo.Wpf/Views/LauncherView.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows.Controls;
+
+namespace GameLib.Demo.Wpf.Views;
+
+public partial class LauncherView : UserControl
+{
+    public LauncherView()
+    {
+        InitializeComponent();
+    }
+}