+++ GameLib.Plugin/GameLib.Plugin.Steam/Model/SteamGame.cs
@@ -1,6 +1,7 @@
 ﻿using Gamelib.Core.Util;
 using GameLib.Core;
 using Microsoft.Win32;
+using System.Drawing;
 
 namespace GameLib.Plugin.Steam.Model;
 
@@ -13,6 +14,7 @@ public class SteamGame : IGame
     public string InstallDir { get; internal set; } = string.Empty;
     public string ExecutablePath { get; internal set; } = string.Empty;
     public string Executable { get; internal set; } = string.Empty;
+    public Icon? ExecutableIcon => PathUtil.GetFileIcon(ExecutablePath);
     public string WorkingDir { get; internal set; } = string.Empty;
     public string LaunchString { get; internal set; } = string.Empty;
     public DateTime InstallDate { get; internal set; } = DateTime.MinValue;