+++ GameLib.Plugin/GameLib.Plugin.Ubisoft/UbisoftLauncher.cs
@@ -36,6 +36,8 @@ public class UbisoftLauncher : ILauncher
 
     public string Executable { get; private set; } = string.Empty;
 
+    public Icon? ExecutableIcon => PathUtil.GetFileIcon(ExecutablePath);
+
     public IEnumerable<IGame> Games { get; private set; } = Enumerable.Empty<IGame>();
 
     public void Refresh(CancellationToken cancellationToken = default)